+++ packages/plugin-autocapture-browser/test/autocapture-plugin/frustration-plugin.test.ts
@@ -539,6 +539,7 @@ describe('frustrationPlugin', () => {
         checkbox.focus(); // This sets document.activeElement to the checkbox
         (window.document as any).dispatchEvent(new Event('selectionchange'));
         document.body.removeChild(checkbox);
+        expect(selectionSpy).not.toHaveBeenCalled();
       });
     });
   });