+++ packages/plugin-page-url-enrichment-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.12](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-url-enrichment-browser@0.5.11...@amplitude/plugin-page-url-enrichment-browser@0.5.12) (2026-01-15)
+
+**Note:** Version bump only for package @amplitude/plugin-page-url-enrichment-browser
+
+
+
+
+
 ## [0.5.11](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-url-enrichment-browser@0.5.10...@amplitude/plugin-page-url-enrichment-browser@0.5.11) (2026-01-14)
 
 **Note:** Version bump only for package @amplitude/plugin-page-url-enrichment-browser