+++ packages/plugin-page-view-tracking-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.9](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.6.8...@amplitude/plugin-page-view-tracking-browser@2.6.9) (2026-01-15)
+
+**Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser
+
+
+
+
+
 ## [2.6.8](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.6.7...@amplitude/plugin-page-view-tracking-browser@2.6.8) (2026-01-14)
 
 **Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser